Revert "Add frontend layer"
