+++ src/main/java/com/inventario/inventario/controller/GuiaController.java
@@ -1,0 +1,7 @@
+package com.inventario.inventario.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class GuiaController {
+}
