+++ src/main/java/com/inventario/inventario/controller/ProductoController.java
@@ -1,7 +1,8 @@
-package com.inventario.inventario.controllers;
+package com.inventario.inventario.controller;
 
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class GuiaController {
+public class ProductoController {
+
 }
