--- src/main/java/com/inventario/inventario/repositories/GuiaRepository.java
@@ -1,4 +1,0 @@
-package com.inventario.inventario.repositories;
-
-public interface GuiaRepository {
-}
