+++ src/main/java/com/inventario/inventario/repository/GuiaRepository.java
@@ -1,0 +1,4 @@
+package com.inventario.inventario.repository;
+
+public interface GuiaRepository {
+}
