+++ src/main/java/com/inventario/inventario/service/GuiaService.java
@@ -1,0 +1,4 @@
+package com.inventario.inventario.service;
+
+public class GuiaService {
+}
