+++ src/main/java/com/inventario/inventario/service/ProductoService.java
@@ -1,0 +1,4 @@
+package com.inventario.inventario.service;
+
+public class ProductoService {
+}
