--- src/main/java/com/inventario/inventario/services/GuiaService.java
@@ -1,4 +1,0 @@
-package com.inventario.inventario.services;
-
-public class GuiaService {
-}
