+++ src/main/resources/application.properties
@@ -1,21 +1,1 @@
 spring.application.name=inventario
-server.port=4000
-
-spring.datasource.url=jdbc:mysql://localhost:3306/inventario?useSSL=false&serverTimezone=America/Santiago
-spring.datasource.username=root
-spring.datasource.password=admin
-spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver
-spring.jpa.data-platform=org.hibernate.dialect.MySQLDialect
-spring.jpa.show-sql=true
-#spring.jpa.hibernate.ddl-auto=create-drop
-logging.level.org.hibernate.SQL=debug
-
-#Para Pruebas
-#spring.datasource.url = jdbc:h2:mem:testbd
-#spring.datasource.driverClassName = org.h2.Driver
-#spring.datasource.username = sa
-#spring.datasource.password = password
-#spring.jpa.show-sql = true
-
-
-
