+++ src/test/java/com/inventario/inventario/InventarioApplicationTests.java
@@ -1,13 +1,13 @@
 package com.inventario.inventario;
 
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
+import org.springframework.boot.test.context.SpringBootTest;
 
-
+@SpringBootTest
 class InventarioApplicationTests {
 
 	@Test
-	public void contextLoads() {
+	void contextLoads() {
 	}
 
 }
